refactor: home/search 코드 품질 및 이미지 최적화
- console.log 제거 (preference/finish)
- dead code 제거: topicroom/search에서 isJoined 중복 분기 삭제
- formatTimeAgo 공통 유틸 추출 (src/lib/utils/formatTimeAgo.ts)
- sentinelRef 선언 순서 버그 수정 (topicroom/search/ResultClient)
- sw.ts fallback URL /~offline → /offline 수정
- SearchBar 뒤로가기 <img> → next/image 전환
- SearchResultWorks 썸네일 sizes="87px" 추가, 별 아이콘 priority 제거
