+++ storix-fe/src/app/sw.ts
@@ -22,7 +22,7 @@ const serwist = new Serwist({
   fallbacks: {
     entries: [
       {
-        url: '/~offline',
+        url: '/offline',
         matcher({ request }) {
           return request.destination === 'document'
         },
