+++ storix-fe/src/lib/utils/formatTimeAgo.ts
@@ -1,0 +1,13 @@
+export function formatTimeAgo(iso?: string | null): string {
+  if (!iso) return ''
+  const t = new Date(iso).getTime()
+  if (Number.isNaN(t)) return ''
+  const diff = Date.now() - t
+  if (diff < 60_000) return '방금 전'
+  const min = Math.floor(diff / 60_000)
+  if (min < 60) return `${min}분 전`
+  const hour = Math.floor(min / 60)
+  if (hour < 24) return `${hour}시간 전`
+  const day = Math.floor(hour / 24)
+  return `${day}일 전`
+}
